refactory: remove internal package
