+++ cmd/api/http/fiber/docs/swagger.yaml
