+++ cmd/api/http/fiber/handler/swagger.go
@@ -3,7 +3,7 @@ package handler
 import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/swagger"
-	_ "github.com/uesleicarvalhoo/sorveteria-tres-estrelas/internal/http/fiber/docs" // Load swagger docs
+	_ "github.com/uesleicarvalhoo/sorveteria-tres-estrelas/cmd/api/http/fiber/docs" // Load swagger docs
 )
 
 func MakeSwaggerRoutes(r fiber.Router) {
