+++ cmd/api/http/fiber/handlers.go
@@ -10,10 +10,10 @@ import (
 	"github.com/gofiber/fiber/v2/middleware/requestid"
 	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/auth"
 	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/cashflow"
+	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/cmd/api/http/fiber/handler"
+	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/cmd/api/http/fiber/middleware"
 	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/healthcheck"
-	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/internal/http/fiber/handler"
-	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/internal/http/fiber/middleware"
-	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/internal/logger"
+	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/logger"
 	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/payments"
 	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/products"
 	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/sales"
