+++ payments/postgres/repository_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
-	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/internal/database"
+	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/database"
 	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/payments"
 	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/payments/postgres"
 	"gorm.io/gorm"
